refactor: move upsert into DonorMetaRepository
